requirements.txt
absl-py==2.3.1
cffi==2.0.0
flatbuffers==25.12.19
mediapipe==0.10.31
numpy==2.2.6
opencv-python==4.12.0.88
pycparser==2.23
python-dotenv==1.2.1
sounddevice==0.5.3
tapo==0.8.8
